fix(ocla): split settlement evidence tests for loc gate
